fix: i64 CTFE 修复——lexer int→int64_t + typeck int→int64_t + codegen %d→%lld
三处 bug 导致 i64 常量表达式被截断为 32 位：
1. lexer.c: lex_number int ival → int64_t ival（9223372036854775807 溢出 int）
2. lexer.c: lex_hex_number (int)(unsigned)(val & 0xFFFFFFFF) → (int64_t)val（截断）
3. typeck.c: fold_expr int l,r → int64_t l,r（二元+一元运算溢出）
4. codegen.c: fprintf %d → i64 时用 %lld（输出截断）

修复后 i64_min = 0 - 9223372036854775807 - 1 = -9223372036854775808 ✅
EXIT=42 ✅（之前 EXIT=232 错误）

macOS gate 验证通过。

+++ compiler/src/typeck/typeck.c
@@ -1713,7 +1713,7 @@ static void fold_expr(struct ASTExpr *e, const char **names, const int *const_va
                 (e->resolved_type->kind == AST_TYPE_I32 || e->resolved_type->kind == AST_TYPE_BOOL ||
                  e->resolved_type->kind == AST_TYPE_U8 || e->resolved_type->kind == AST_TYPE_U32 ||
                  e->resolved_type->kind == AST_TYPE_I64 || e->resolved_type->kind == AST_TYPE_USIZE || e->resolved_type->kind == AST_TYPE_ISIZE)) {
-                int l = e->value.binop.left->const_folded_val, r = e->value.binop.right->const_folded_val;
+                int64_t l = e->value.binop.left->const_folded_val, r = e->value.binop.right->const_folded_val;
                 switch (e->kind) {
                     case AST_EXPR_ADD: e->const_folded_val = l + r; break;
                     case AST_EXPR_SUB: e->const_folded_val = l - r; break;
@@ -1746,7 +1746,7 @@ static void fold_expr(struct ASTExpr *e, const char **names, const int *const_va
                 (e->resolved_type->kind == AST_TYPE_I32 || e->resolved_type->kind == AST_TYPE_BOOL ||
                  e->resolved_type->kind == AST_TYPE_U8 || e->resolved_type->kind == AST_TYPE_U32 ||
                  e->resolved_type->kind == AST_TYPE_I64 || e->resolved_type->kind == AST_TYPE_USIZE || e->resolved_type->kind == AST_TYPE_ISIZE)) {
-                int o = e->value.unary.operand->const_folded_val;
+                int64_t o = e->value.unary.operand->const_folded_val;
                 if (e->kind == AST_EXPR_NEG) e->const_folded_val = -o;
                 else if (e->kind == AST_EXPR_BITNOT) e->const_folded_val = ~o;
                 else e->const_folded_val = !o ? 1 : 0;
